+++ shared/src/commonMain/kotlin/com/example/digisim/LogicGates/Nor.kt
@@ -22,7 +22,7 @@ import com.example.digisim.DrawingLogic.calculateInputPinPositions
      override var height: Float by mutableStateOf(60f)
      override var inputCount: Int by mutableStateOf(initialInputCount)
      override var outputCount: Int by mutableStateOf(initialOutputCount)
-     override var delay: Int? = null
+     override var delay: Float? = null
 
      override val componentType = ComponentType.NOR
